requirements.txt
fastapi
uvicorn
sqlalchemy
transformers
torch
spacy
vaderSentiment
geotext
slowapi
praw
telethon
requests
feedparser
pydantic
python-dotenv
langdetect 